modified:   README.md 	modified:   setup.py 	modified:   svdb/__main__.py 	modified:   svdb/export_module.py

+++ svdb/__main__.py
@@ -52,7 +52,7 @@ def main():
         parser.add_argument('--bnd_distance', type=int,default= 10000,help="the maximum distance between two similar precise breakpoints(default = 10000)")
         parser.add_argument('--overlap', type=float, default = 0.6,help="the overlap required to merge two events(0 means anything that touches will be merged, 1 means that two events must be identical to be merged), default = 0.6")
         parser.add_argument('--DBSCAN'       , help="use dbscan to cluster the variants, only available for the sqlite db", required=False, action="store_true")
-        parser.add_argument('--epsilon'       ,type=int, default = 500, help="used together with --DBSCAN; sets the epsilon paramter(default = 500)", required=False)
+        parser.add_argument('--epsilon'       ,type=float, default = 500, help="used together with --DBSCAN; sets the epsilon paramter(default = 500)", required=False)
         parser.add_argument('--min_pts'       ,type=int, default = 2, help="used together with 1--DBSCAN; sets the min_pts parameter(default = 2)", required=False)  
         parser.add_argument('--memory'       , help="load the database into memory: increases the memory requirements, but lowers the time consumption(may only be used with sqdb)", required=False, action="store_true")        
         
@@ -89,7 +89,7 @@ def main():
         parser.add_argument('--bnd_distance', type=int,default= 2500,help="the maximum distance between two similar precise breakpoints(default = 2500)")
         parser.add_argument('--overlap', type=float, default = 0.8,help="the overlap required to merge two events(0 means anything that touches will be merged, 1 means that two events must be identical to be merged), default = 0.8")
         parser.add_argument('--DBSCAN'       , help="use dbscan to cluster the variants", required=False, action="store_true")
-        parser.add_argument('--epsilon'       ,type=int, default = 500, help="used together with --DBSCAN; sets the epsilon paramter(default = 500)", required=False)
+        parser.add_argument('--epsilon'       ,type=float, default = 500, help="used together with --DBSCAN; sets the epsilon paramter(default = 500)", required=False)
         parser.add_argument('--min_pts'       ,type=int, default = 2, help="used together with 1--DBSCAN; sets the min_pts parameter(default = 2)", required=False)              
         parser.add_argument('--prefix', type=str,default="SVDB" ,help="the prefix of the output file, default = same as input")
         parser.add_argument('--memory'       , help="load the database into memory: increases the memory requirements, but lowers the time consumption", required=False, action="store_true")
